Implement mouse pointing and panning
To do this some code has been moved around. Renderer now tracks whether
a redraw is needed, and now has mouse callbacks in viewport space from
the Window. These are converted into a mouse ray in world coordinates
(orthographic for now so fairly straightforward), via the Viewpoint
class (which now provides glOrtho parameters for the RendererOpenGL to
use).

Renderer passes the events on to a ControlMode class (first the active
mode, then the navigate mode). This abstracts mouse mode operations, and
for now we activate a NavigateMode which takes care of zooming and
panning using these mouse events, as well as doing some OpenGL specific
rendering so we can see a point at the mouse position.

Signed-off-by: James Hogan <james@albanarts.com>