Add ClothoidChain
Add a simple Clothoid chain implementation using a std::list. It starts
empty, and append() will return a pointer to a new Clothoid starting
where the last one finished, which can then be changed.

Also available are length() to get the total length of the chain, and
positionAtLength() to get the position of the curve at a certain length
from the start.

Signed-off-by: James Hogan <james@albanarts.com>