Implement TrackSection with clothoids
Add a clothoid chain to TrackSection, implement some clothoid
interpolation, and rendering of a line based train track along the
clothoid chain.

The interpolation is pending some TrackMode improvements, but it
attempts to interpolate between two nodes with certain directions and
curvatures using 2 curves with a straight in between, with 4 transition
curves. The curves are of fixed curvature at the moment. In the event
that the curve needs a full loop, the curve and transition curves on
either side are switched to a pair of clothoid curves, potentially
changing the length of the straight. This ends up using Newton's method
to find the appropriate curvature rate to line up with the straight.

There is certainly room for improvement, but it fundamentally seems to
work acceptably.

Signed-off-by: James Hogan <james@albanarts.com>