TrackMode: Fix end node z depending on zoom
The end node z coordinate was based on the mouse ray start coordinate,
which moves with zoom. This resulted in a zoomed out creation of a track
with the end node unclickable and invisible.

For now cast to Vec2f and Vec3f to strip off the z coordinate, as is
done for the start node.

Signed-off-by: James Hogan <james@albanarts.com>